Adjusts withdrawal schedule to Wednesday 00:00 UTC
Updates the scheduled withdrawals to occur on Wednesdays at midnight UTC instead of Tuesdays at 6 PM UTC.
This also adjusts the monthly withdrawal date to the 8th of the month at midnight UTC.
The start date for new wallets is set to October 8th, 2025.

This change aligns with a desired adjustment to the timing of withdrawals.

Additionally, this commit includes:
- A downgrade of better-sqlite3 dependency to resolve compatibility issues.
- A new `start.sh` script designed for Gandi production environment deployment to initialize dependencies automatically.